[tserver] fixed build
git-svn-id: https://tcoffee.googlecode.com/svn/tserver/trunk@683 cb409b52-f269-45c4-4a06-1bd0eacd059b

+++ build/build.sh
@@ -239,6 +239,11 @@ function test_server_local()
 	# start a play instance and invoke tests
 	cd $SANDBOX/web
 	./$PLAY_VER/play auto-test tserver 
+	count=`ls ./tserver/test-result/*.html | wc -l`
+	if [[ $count==0 ]]; then
+	  cat ./tserver/test-result/application.log
+	  exit 1;
+	fi 
 
 	# publish the tests
 	rm -rf $WORKSPACE/test-result
